pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name        = "sshcrack"
version     = "1.0.0"
description = "Multi-process SSH private key passphrase cracker — OpenSSH + PuTTY PPK"
readme      = "README.md"
license     = { text = "MIT" }
requires-python = ">=3.8"

authors = [
    { name = "Bhanu Guragain" }
]

keywords = [
    "ssh", "passphrase", "cracker", "pentest",
    "red-team", "security", "openssh", "putty", "ppk"
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "cryptography[bcrypt]>=38.0.0",
]

[project.optional-dependencies]
verify        = ["paramiko>=3.0.0"]
ppk-v3        = ["argon2-cffi>=21.0.0"]
gpu-cuda      = ["pycuda>=2022.1"]
gpu-opencl    = ["pyopencl>=2022.1"]
distributed   = ["pyzmq>=25.0.0"]
all           = [
    "paramiko>=3.0.0",
    "argon2-cffi>=21.0.0",
    "pyzmq>=25.0.0",
]
dev           = [
    "pytest>=7.4",
    "pytest-cov>=4.0",
    "coverage>=7.5",
    "mypy>=1.10",
    "ruff==0.4.4",
    "black==24.4.2",
    "isort==5.13.2",
    "bandit[toml]>=1.7.9",
    "pip-audit>=2.7",
    "pre-commit>=3.7",
]

[project.scripts]
sshcrack = "sshcrack.cli:main"

[project.urls]
Homepage   = "https://github.com/GANGAOps/SSHCrack"
Repository = "https://github.com/GANGAOps/SSHCrack"
Issues     = "https://github.com/GANGAOps/SSHCrack/issues"

# ── setuptools config ──────────────────────────────────────────────────────────

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
sshcrack = ["py.typed"]

# ── pytest ─────────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths    = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts      = "-v --tb=short"

# ── coverage ───────────────────────────────────────────────────────────────────

[tool.coverage.run]
source  = ["sshcrack"]
omit    = ["sshcrack/cpu/__init__.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true

# ── ruff (linter + formatter) ─────────────────────────────────────────────────

[tool.ruff]
line-length    = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "N",    # pep8-naming
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",   # line too long (black handles this)
    "B904",   # raise from None in except (intentional)
    "SIM108", # ternary operator (readability preference)
    "N803",   # argument name should be lowercase (matrix params OK)
    "N806",   # variable in function should be lowercase
]
fixable = ["I", "UP", "RUF"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S603", "S607"]   # assert OK in tests

[tool.ruff.format]
quote-style             = "double"
indent-style            = "space"
skip-magic-trailing-comma = false

# ── mypy ───────────────────────────────────────────────────────────────────────

[tool.mypy]
python_version         = "3.12"
strict                 = false
ignore_missing_imports = true
warn_return_any        = true
warn_unused_configs    = true
show_error_codes       = true
pretty                 = true
# Gradual typing — enforce on core modules, warn on others
[[tool.mypy.overrides]]
module = ["sshcrack.parser", "sshcrack.engine", "sshcrack.session"]
strict = true
[[tool.mypy.overrides]]
module = ["pycuda.*", "pyopencl.*", "zmq.*", "paramiko.*"]
ignore_missing_imports = true

# ── bandit (security scanner) ─────────────────────────────────────────────────

[tool.bandit]
exclude_dirs = ["tests"]
skips        = ["B404", "B603"]   # subprocess usage is intentional

# ── black ──────────────────────────────────────────────────────────────────────

[tool.black]
line-length    = 100
target-version = ["py311", "py312"]
skip-string-normalization = false

# ── isort ──────────────────────────────────────────────────────────────────────

[tool.isort]
profile                   = "black"
line_length               = 100
known_first_party         = ["sshcrack"]
known_third_party         = ["cryptography", "numpy", "zmq", "paramiko", "pyopencl", "pycuda"]
force_sort_within_sections = true
